Set Fluentd dependency
Because this plugin supports Fluentd v0.14.x only, does not support v0.12.x.
Fluentd v1 will be compatible with v0.14.x.
Fluentd v2 may not be compatible with v1 and v0.14.x few years later.

This is useful for users to understand that this plugin supports
Fluentd v0.14.x only.